close_range: use unsigned int in protected-fd range guard
The kernel's close_range syscall takes (unsigned int first, unsigned
int last) and walks the fd table whenever first <= last as unsigned.
sanitise_close_range's protected-fd truncation passed the bounds
through lowest_protected_fd_in_range() as signed int, so an a2 of
(unsigned long)-1 -- gen_arg_fd's exhaustion fallback -- arrived as
int -1, tripped the helper's hi < lo guard, and returned -1 (no
protected fd in range).  The kernel then saw close_range(a1,
0xFFFFFFFF) and walked the entire fd table, closing the kcov PC and
CMP fds at KCOV_FD_HIGH_BASE.  Every later kcov_pc_enable /
kcov_pc_disable on that child then returned EBADF, silently
under-counting coverage for the rest of the run.

The same signed-int gap hid the close in the diag: the chronicle
check kcov_chronicle_slot_touched_protected() cast s->a1/a2 to int
before calling the helper, so the per-EBADF latch reported
"closer=close_range" without the /prot suffix that should mark a
fuzz-syscall as the cause.

Match the kernel ABI throughout the protected-fd path:

  - lowest_protected_fd_in_range() now takes (unsigned int lo,
    unsigned int hi) and compares the protected-fd values against
    them as unsigned int.  All currently-protected fds (STDERR_FILENO,
    the stderr capture memfd, kcov.fd, kcov.cmp_fd) fit in int, so
    the return type stays int with -1 for "no protected fd in
    range".

  - sanitise_close_range gates the truncation on (unsigned int) a1
    <= (unsigned int) a2 and casts the bounds to unsigned int before
    handing them to the helper, so the kernel-side walk is bounded
    by lowest_protected - 1 (or inverted to EINVAL) regardless of
    whether gen_arg_fd handed back -1.

  - kcov_chronicle_slot_touched_protected() does the same cast on
    the chronicle slot's saved a1/a2, so the diag /prot suffix now
    correctly fires for any close_range whose range covered a
    protected fd.

Gates: CFLAGS="-g -O2 -D_FORTIFY_SOURCE=2 -Werror" make -j16 clean &&
make -j16 clean.  scripts/check-static.sh: 27/27 pass.  ./trinity
--dry-run -N 200 -C2 --no-cgroup: rc=0.  ./trinity --dry-run -N 5000
-C2 --no-cgroup -c close_range -c close: rc=0, no
pc_enable=EBADF/pc_disable=EBADF/close_range_protect diag emitted.

+++ syscalls/close_range.c
@@ -230,10 +230,19 @@ static void sanitise_close_range(struct syscallrecord *rec)
 	 * close (lowest protected fd <= rec->a1) we collapse the range to
 	 * a no-op (close_range(N, N - 1) returns -EINVAL) rather than
 	 * issuing a side-effect-free syscall on an unrelated low slot.
+	 *
+	 * Compare and pass bounds as unsigned int: the kernel's
+	 * close_range ABI takes (unsigned int first, unsigned int last),
+	 * and gen_arg_fd's exhaustion fallback hands back (unsigned
+	 * long)-1, which the kernel sees as 0xFFFFFFFF.  A signed-int
+	 * compare here would treat that as a negative "hi", skip the
+	 * truncation, and let the kernel walk [a1, 0xFFFFFFFF] -- the
+	 * exact path that closes the kcov fd at KCOV_FD_HIGH_BASE.
 	 */
-	if (rec->a1 <= rec->a2) {
-		int lowest = lowest_protected_fd_in_range((int) rec->a1,
-							  (int) rec->a2);
+	if ((unsigned int) rec->a1 <= (unsigned int) rec->a2) {
+		int lowest = lowest_protected_fd_in_range(
+				(unsigned int) rec->a1,
+				(unsigned int) rec->a2);
 		if (lowest >= 0) {
 			if ((unsigned int) lowest <= (unsigned int) rec->a1) {
 				/* Whole range was at or above the lowest
